ArchArchivesBorrowStatisticsMapper.java ArchArchivesBorrowStatisticsMapper.xml ArchArchivesBorrowStatisticsServiceImpl.java ArchArchivesQuantityStatisticsMapper.java ArchArchivesQuantityStatisticsMapper.xml ArchArchivesQuantityStatisticsServiceImpl.java ArchDiseaseArchivesAppraisalMapper.java ArchDiseaseArchivesAppraisalMapper.xml ArchDiseaseArchivesAppraisalServiceImpl.java ArchSearchArchivesQuantityStatisticsMapper.xml ArchUsingPurposeStatisticsMapper.java IInxSearchArchivesQuantityAndUsingPurposeStatisticsService.java InxSearchArchivesQuantityStatisticsAndUsingPurposeController.java InxSearchArchivesQuantityStatisticsAndUsingPurposeMapper.java InxSearchArchivesQuantityStatisticsAndUsingPurposeMapper.xml InxSearchArchivesQuantityStatisticsAndUsingPurposeServiceImpl.java InxSearchArchivesQuantityStatisticsVo.java InxUsingPurposeStatisticsQueryDto.java files

+++ archives-manage-dao/src/main/java/com/yintu/rixing/archives/ArchArchivesBorrowStatisticsMapper.java
@@ -27,5 +27,5 @@ public interface ArchArchivesBorrowStatisticsMapper {
      * @param endDate           结束日期
      * @return ..
      */
-    Map<String, BigDecimal> selectArchivesBorrowStatisticsData(Integer archivesLibraryId, String tableName, Short borrowType, Integer departmentId, Date startDate, Date endDate);
+    Map<String, Long> selectArchivesBorrowStatisticsData(Integer archivesLibraryId, String tableName, Short borrowType, Integer departmentId, Date startDate, Date endDate);
 }
